pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repodoctor"
version = "0.1.0"
description = "Automated technical documentation generator for code repositories using LLM analysis"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Alexander Orlov", email = "pipy@orlov.ai"}
]
maintainers = [
    {name = "Alexander Orlov", email = "pipy@orlov.ai"}
]
keywords = [
    "documentation", 
    "generator", 
    "llm", 
    "code-analysis", 
    "markdown", 
    "docusaurus",
    "openrouter"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Utilities"
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/AlexanderOrlov/repodoctor"
Documentation = "https://github.com/AlexanderOrlov/repodoctor#readme"
Repository = "https://github.com/AlexanderOrlov/repodoctor.git"
Issues = "https://github.com/AlexanderOrlov/repodoctor/issues"
Changelog = "https://github.com/AlexanderOrlov/repodoctor/blob/main/CHANGELOG.md"

[project.scripts]
repodoctor = "repodoctor.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["repodoctor*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
repodoctor = ["*.md", "*.txt", "*.yml", "*.yaml"]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=repodoctor",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80"
]
